requirements.txt
APScheduler==3.6.3
astroid==2.9.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
Deprecated==1.2.13
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.3
platformdirs==2.4.1
pylint==2.12.2
pyparsing==3.0.6
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.1.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
tornado==6.1
tweepy==4.4.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
wrapt==1.13.3
